Split iOS progress snapshot factory tests
